Improved compose navigation code structure
