+++ samples/navigation-compose/src/main/java/com/bumble/appyx/sample/navigtion/compose/ComposeNavigationContainerNode.kt
@@ -1,0 +1,62 @@
+package com.bumble.appyx.sample.navigtion.compose
+
+import android.os.Parcelable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.bumble.appyx.core.composable.Children
+import com.bumble.appyx.core.modality.BuildContext
+import com.bumble.appyx.core.node.Node
+import com.bumble.appyx.core.node.ParentNode
+import com.bumble.appyx.core.node.node
+import com.bumble.appyx.routingsource.backstack.BackStack
+import kotlinx.parcelize.Parcelize
+
+class ComposeNavigationContainerNode(
+    buildContext: BuildContext,
+    private val onGoogleNavigationClick: () -> Unit,
+    private val backStack: BackStack<Routing> = BackStack(
+        initialElement = Routing.Main,
+        savedStateMap = buildContext.savedStateMap,
+    )
+) : ParentNode<ComposeNavigationContainerNode.Routing>(
+    routingSource = backStack,
+    buildContext = buildContext,
+) {
+
+    sealed class Routing : Parcelable {
+        @Parcelize
+        object Main : Routing()
+    }
+
+    override fun resolve(routing: Routing, buildContext: BuildContext): Node =
+        when (routing) {
+            is Routing.Main -> node(buildContext) {
+                Column(
+                    modifier = Modifier.fillMaxSize(),
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text("Appyx screen")
+                    Button(onClick = { onGoogleNavigationClick() }) {
+                        Text("Navigate to Google")
+                    }
+                }
+            }
+        }
+
+    @Composable
+    override fun View(modifier: Modifier) {
+        Children(
+            modifier = modifier.fillMaxSize(),
+            routingSource = backStack
+        ) {
+            children<Routing> { child ->
+                child()
+            }
+        }
+    }
+}
